+++ src/api/java/blusunrize/immersiveengineering/api/Lib.java
@@ -75,6 +75,10 @@ public class Lib
 	public static final String GUIID_ToolboxBlock = "toolboxblock";
 	public static final String GUIID_RedstoneConnector = "redstoneconnector";
 	public static final String GUIID_RedstoneProbe = "redstoneprobe";
+	public static final String GUIID_RedstoneStateCell = "redstonestatecell";
+	public static final String GUIID_RedstoneTimer = "redstonetimer";
+	public static final String GUIID_RedstoneSwitchboard = "redstoneswitchboard";
+	public static final String GUIID_Siren = "siren";
 	//Items
 	public static final String GUIID_Revolver = "revolver";
 	public static final String GUIID_Toolbox = "toolbox";