+++ src/datagen/java/blusunrize/immersiveengineering/data/tags/IEBlockTags.java
@@ -40,7 +40,6 @@ import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Tiers;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.SlabBlock;
 import net.neoforged.neoforge.common.Tags;
 import net.neoforged.neoforge.common.data.BlockTagsProvider;
 import net.neoforged.neoforge.common.data.ExistingFileHelper;
@@ -332,6 +331,8 @@ public class IEBlockTags extends BlockTagsProvider
 		}
 		//Razor wire should only be broken by wirecutters, and its addition is accidental
 		tag.remove(MetalDevices.RAZOR_WIRE.get());
+		// switchboard doesn't extend ConnectorBlock
+		tag.add(Connectors.REDSTONE_SWITCHBOARD.get());
 	}
 
 	private void registerRockcutterMineable()
@@ -630,6 +631,10 @@ public class IEBlockTags extends BlockTagsProvider
 				Connectors.CONNECTOR_REDSTONE,
 				Connectors.CONNECTOR_PROBE,
 				Connectors.CONNECTOR_BUNDLED,
+				Connectors.REDSTONE_STATE_CELL,
+				Connectors.REDSTONE_TIMER,
+				Connectors.REDSTONE_SWITCHBOARD,
+				Connectors.SIREN,
 				Connectors.FEEDTHROUGH
 		);
 		registerMineable(tag, Metals.SHEETMETAL);