+++ src/main/java/blusunrize/immersiveengineering/client/ClientProxy.java
@@ -45,14 +45,12 @@ import blusunrize.immersiveengineering.client.utils.BasicClientProperties;
 import blusunrize.immersiveengineering.client.utils.VertexBufferHolder;
 import blusunrize.immersiveengineering.common.CommonProxy;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.ISoundBE;
-import blusunrize.immersiveengineering.common.blocks.metal.ConnectorProbeBlockEntity;
-import blusunrize.immersiveengineering.common.blocks.metal.ConnectorRedstoneBlockEntity;
+import blusunrize.immersiveengineering.common.blocks.metal.*;
 import blusunrize.immersiveengineering.common.blocks.wooden.MachineInterfaceBlockEntity;
 import blusunrize.immersiveengineering.common.config.IEClientConfig;
 import blusunrize.immersiveengineering.common.entities.SkylineHookEntity;
 import blusunrize.immersiveengineering.common.gui.IEBaseContainerOld;
 import blusunrize.immersiveengineering.common.register.*;
-import blusunrize.immersiveengineering.common.register.IEBannerPatterns.BannerEntry;
 import blusunrize.immersiveengineering.common.register.IEMenuTypes.ArgContainer;
 import blusunrize.immersiveengineering.common.util.sound.IEBlockEntitySound;
 import blusunrize.immersiveengineering.common.util.sound.SkyhookSound;
@@ -137,6 +135,8 @@ public class ClientProxy extends CommonProxy
 		IEOBJCallbacks.register(rl("post"), PostCallbacks.INSTANCE);
 		IEOBJCallbacks.register(rl("razor_wire"), RazorWireCallbacks.INSTANCE);
 		IEOBJCallbacks.register(rl("connector_rs"), RSConnectorCallbacks.INSTANCE);
+		IEOBJCallbacks.register(rl("timer"), TimerCallbacks.INSTANCE);
+		IEOBJCallbacks.register(rl("siren"), SirenCallbacks.INSTANCE);
 		IEOBJCallbacks.register(rl("structural_arm"), StructuralArmCallbacks.INSTANCE);
 		IEOBJCallbacks.register(rl("structural_connector"), StructuralConnectorCallbacks.INSTANCE);
 		IEOBJCallbacks.register(rl("turret"), TurretCallbacks.INSTANCE);
@@ -370,6 +370,18 @@ public class ClientProxy extends CommonProxy
 
 		if(guiId.equals(Lib.GUIID_RedstoneProbe)&&tileEntity instanceof ConnectorProbeBlockEntity)
 			Minecraft.getInstance().setScreen(new RedstoneProbeScreen((ConnectorProbeBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));
+
+		if(guiId.equals(Lib.GUIID_RedstoneStateCell)&&tileEntity instanceof RedstoneStateCellBlockEntity)
+			Minecraft.getInstance().setScreen(new RedstoneStateCellScreen((RedstoneStateCellBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));
+
+		if(guiId.equals(Lib.GUIID_RedstoneTimer)&&tileEntity instanceof RedstoneTimerBlockEntity)
+			Minecraft.getInstance().setScreen(new RedstoneTimerScreen((RedstoneTimerBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));
+
+		if(guiId.equals(Lib.GUIID_RedstoneSwitchboard)&&tileEntity instanceof RedstoneSwitchboardBlockEntity)
+			Minecraft.getInstance().setScreen(new RedstoneSwitchboardScreen((RedstoneSwitchboardBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));
+
+		if(guiId.equals(Lib.GUIID_Siren)&&tileEntity instanceof SirenBlockEntity)
+			Minecraft.getInstance().setScreen(new SirenScreen((SirenBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));
 
 		if(guiId.equals(Lib.GUIID_MachineInterface)&&tileEntity instanceof MachineInterfaceBlockEntity)
 			Minecraft.getInstance().setScreen(new MachineInterfaceScreen((MachineInterfaceBlockEntity)tileEntity, tileEntity.getBlockState().getBlock().getName()));