+++ src/main/java/blusunrize/immersiveengineering/client/gui/RedstoneConnectorScreen.java
@@ -81,7 +81,7 @@ public class RedstoneConnectorScreen extends ClientBlockEntityScreen<ConnectorRe
 	{
 		ArrayList<Component> tooltip = new ArrayList<>();
 
-		if(buttonInOut.isHoveredOrFocused())
+		if(buttonInOut.isHovered())
 		{
 			tooltip.add(Component.translatable(Lib.GUI_CONFIG+"redstone_iomode"));
 			tooltip.add(TextUtils.applyFormat(
@@ -91,7 +91,7 @@ public class RedstoneConnectorScreen extends ClientBlockEntityScreen<ConnectorRe
 		}
 
 		for(int i = 0; i < colorButtons.length; i++)
-			if(colorButtons[i].isHoveredOrFocused())
+			if(colorButtons[i].isHovered())
 			{
 				tooltip.add(Component.translatable(Lib.GUI_CONFIG+"redstone_color"));
 				tooltip.add(TextUtils.applyFormat(
@@ -120,7 +120,7 @@ public class RedstoneConnectorScreen extends ClientBlockEntityScreen<ConnectorRe
 			GuiButtonBoolean[] buttons, int posX, int posY, Supplier<Boolean> active, DyeColor color, Consumer<GuiButtonBoolean> onClick
 	)
 	{
-		return new GuiButtonBoolean(posX, posY, 12, 12, "", active,
+		return new GuiButtonBoolean(posX, posY, 12, 12, Component.empty(), active,
 				TEXTURE, 194, 0, 1,
 				btn -> {
 					if(btn.getNextState())