+++ src/main/java/blusunrize/immersiveengineering/client/gui/RedstoneProbeScreen.java
@@ -82,7 +82,7 @@ public class RedstoneProbeScreen extends ClientBlockEntityScreen<ConnectorProbeB
 
 		ArrayList<Component> tooltip = new ArrayList<>();
 		for(int i = 0; i < colorButtonsSend.length; i++)
-			if(colorButtonsSend[i].isHoveredOrFocused()||colorButtonsReceive[i].isHoveredOrFocused())
+			if(colorButtonsSend[i].isHovered()||colorButtonsReceive[i].isHovered())
 			{
 				tooltip.add(Component.translatable(Lib.GUI_CONFIG+"redstone_color"));
 				tooltip.add(TextUtils.applyFormat(