+++ src/main/java/blusunrize/immersiveengineering/common/util/IESounds.java
@@ -59,6 +59,9 @@ public class IESounds
 	public static final Holder<SoundEvent> process2Lift = registerSound("process_2_lift");
 	public static final Holder<SoundEvent> mill_creaking = registerSound("mill_creaking");
 	public static final Holder<SoundEvent> electromagnet = registerSound("electromagnet");
+	public static final Holder<SoundEvent> siren = registerSound("siren");
+	public static final Holder<SoundEvent> klaxon = registerSound("klaxon");
+	public static final Holder<SoundEvent> buzzer = registerSound("buzzer");
 
 
 	public static void init(IEventBus modBus)