+++ changelog.md
@@ -1,8 +1,22 @@
-##### Version 1.20.4-11.5.1-pre
+##### Version 1.20.4-11.6.0-pre
+- "The Redstone Update!"
+    - Add the redstone state cell for storing signals (BluSunrize)
+    - Add the redstone timer for emitting regular signals (BluSunrize)
+    - Add the redstone switchboard for remapping redstone signals between networks (BluSunrize)
+    - Add the warning siren to make loud noises and send villagers to their homes (BluSunrize)
+    - Move redstone components to a separate category of the manual (BluSunrize)
+- Add recipe for crushing slag bricks into slag gravel (BluSunrize)
+- Add recipes for wetting vanilla concrete powder in the bottling machine (BluSunrize)
+- Add progress tooltip to items in the arc furnace (BluSunrize)
+- Change recipe for catwalk stairs to be cheaper (BluSunrize)
 - Fix waterlogged posts not dropping items (BluSunrize)
 - Fix treated windows and feedthrough insulators letting rain through (BluSunrize, voidsong-dragonfly)
 - Fix crates not syncing their names to the client (BluSunrize)
 - Fix shader based particles breaking on servers (BluSunrize)
+- Fix various button tooltips sticking around after being clicked (BluSunrize)
+- Fix treated trapdoors erroneously having metal properties (BluSunrize)
+- Fix arms on posts dropping blocks when broken (BluSunrize)
+- Fix crate minecarts crashing when inserted into with a hopper (BluSunrize)
 
 ##### Version 1.20.4-11.5.0-179
 - Add warning signs (BluSunrize)